+++ lib/commonjs/ReactNativeZoomableView.js
@@ -3,7 +3,14 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.default = exports.ReactNativeZoomableViewContext = exports.ReactNativeZoomableView = void 0;
+exports.ReactNativeZoomableView = void 0;
+Object.defineProperty(exports, "ReactNativeZoomableViewContext", {
+  enumerable: true,
+  get: function () {
+    return _ReactNativeZoomableViewContext.ReactNativeZoomableViewContext;
+  }
+});
+exports.default = void 0;
 var _lodash = require("lodash");
 var _react = _interopRequireWildcard(require("react"));
 var _reactNative = require("react-native");
@@ -19,9 +26,9 @@ var _getNextZoomStep = require("./helper/getNextZoomStep");
 var _useDebugPoints = require("./hooks/useDebugPoints");
 var _useLatestCallback = require("./hooks/useLatestCallback");
 var _useZoomSubject = require("./hooks/useZoomSubject");
+var _ReactNativeZoomableViewContext = require("./ReactNativeZoomableViewContext");
 var _jsxRuntime = require("react/jsx-runtime");
 function _interopRequireWildcard(e, t) { if ("function" == typeof WeakMap) var r = new WeakMap(), n = new WeakMap(); return (_interopRequireWildcard = function (e, t) { if (!t && e && e.__esModule) return e; var o, i, f = { __proto__: null, default: e }; if (null === e || "object" != typeof e && "function" != typeof e) return f; if (o = t ? n : r) { if (o.has(e)) return o.get(e); o.set(e, f); } for (const t in e) "default" !== t && {}.hasOwnProperty.call(e, t) && ((i = (o = Object.defineProperty) && Object.getOwnPropertyDescriptor(e, t)) && (i.get || i.set) ? o(f, t, i) : f[t] = e[t]); return f; })(e, t); }
-const ReactNativeZoomableViewContext = exports.ReactNativeZoomableViewContext = /*#__PURE__*/_react.default.createContext(undefined);
 const ReactNativeZoomableView = (props, ref) => {
   const {
     wrapperRef: zoomSubjectWrapperRef,
@@ -742,7 +749,7 @@ const ReactNativeZoomableView = (props, ref) => {
   }).onFinalize(() => {
     firstTouch.value = undefined;
   });
-  return /*#__PURE__*/(0, _jsxRuntime.jsx)(ReactNativeZoomableViewContext.Provider, {
+  return /*#__PURE__*/(0, _jsxRuntime.jsx)(_ReactNativeZoomableViewContext.ReactNativeZoomableViewContext.Provider, {
     value: {
       zoom,
       inverseZoom,
