+++ lib/commonjs/ReactNativeZoomableViewContext.js
@@ -1,0 +1,9 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.ReactNativeZoomableViewContext = void 0;
+var _react = require("react");
+const ReactNativeZoomableViewContext = exports.ReactNativeZoomableViewContext = /*#__PURE__*/(0, _react.createContext)(undefined);
+//# sourceMappingURL=ReactNativeZoomableViewContext.js.map
