+++ lib/commonjs/components/FixedSize.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", {
 exports.default = exports.FixedSize = void 0;
 var _react = _interopRequireDefault(require("react"));
 var _reactNativeReanimated = _interopRequireDefault(require("react-native-reanimated"));
-var _ReactNativeZoomableView = require("../ReactNativeZoomableView");
+var _ReactNativeZoomableViewContext = require("../ReactNativeZoomableViewContext");
 var _jsxRuntime = require("react/jsx-runtime");
 function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
 /**
@@ -27,7 +27,7 @@ const FixedSize = ({
   top,
   children
 }) => {
-  const context = _react.default.useContext(_ReactNativeZoomableView.ReactNativeZoomableViewContext);
+  const context = _react.default.useContext(_ReactNativeZoomableViewContext.ReactNativeZoomableViewContext);
   return /*#__PURE__*/(0, _jsxRuntime.jsx)(_reactNativeReanimated.default.View, {
     style: [context?.inverseZoomStyle, {
       width: 1,
