+++ lib/module/ReactNativeZoomableView.js
@@ -1,7 +1,8 @@
 "use strict";
 
 import { debounce, defaults } from 'lodash';
-import React, { useImperativeHandle, useLayoutEffect, useMemo, useRef, useState } from 'react';
+import React from 'react';
+import { useImperativeHandle, useLayoutEffect, useMemo, useRef, useState } from 'react';
 import { StyleSheet, View } from 'react-native';
 import { Gesture, GestureDetector, GestureHandlerRootView } from 'react-native-gesture-handler';
 import Animated, { cancelAnimation, makeMutable, runOnJS, useAnimatedReaction, useAnimatedStyle, useDerivedValue, useSharedValue, withTiming } from 'react-native-reanimated';
@@ -15,8 +16,8 @@ import { getNextZoomStep } from './helper/getNextZoomStep';
 import { useDebugPoints } from './hooks/useDebugPoints';
 import { useLatestCallback } from './hooks/useLatestCallback';
 import { useZoomSubject } from './hooks/useZoomSubject';
+import { ReactNativeZoomableViewContext } from './ReactNativeZoomableViewContext';
 import { jsx as _jsx, jsxs as _jsxs } from "react/jsx-runtime";
-const ReactNativeZoomableViewContext = /*#__PURE__*/React.createContext(undefined);
 const ReactNativeZoomableView = (props, ref) => {
   const {
     wrapperRef: zoomSubjectWrapperRef,
