+++ lib/module/ReactNativeZoomableViewContext.js
@@ -1,0 +1,5 @@
+"use strict";
+
+import { createContext } from 'react';
+export const ReactNativeZoomableViewContext = /*#__PURE__*/createContext(undefined);
+//# sourceMappingURL=ReactNativeZoomableViewContext.js.map
